requirements.txt
# Core data
pandas
sqlalchemy
psycopg2-binary
openpyxl
python-calamine

# Config / utils
pyyaml

# Orchestration
prefect

# LLM / AI (Tier 3 CRS classifier)
langchain-openai
langchain-core

# UI
streamlit

# HTTP clients
httpx
requests

# Auth
bcrypt

# Secrets management
python-dotenv>=1.0.0
